Split VTT.js into: vttparser, vttcue (polyfill), vttregion (polyfill), vttrenderer. Updated mixin and renderer to load files async only when needed.
