+++ src/js/controller/setup-steps.js
@@ -24,11 +24,21 @@ define([
                 method: _loadBase64Polyfill,
                 depends: []
             },
+            LOAD_VTTCUE_POLYFILL : {
+                method: _loadVTTCuePolyfill,
+                depends: []
+            },
+            LOAD_VTTREGION_POLYFILL : {
+                method: _loadVTTRegionPolyfill,
+                depends: []
+            },
             LOADED_POLYFILLS : {
                 method: _loadedPolyfills,
                 depends: [
                     'LOAD_PROMISE_POLYFILL',
-                    'LOAD_BASE64_POLYFILL'
+                    'LOAD_BASE64_POLYFILL',
+                    'LOAD_VTTCUE_POLYFILL',
+                    'LOAD_VTTREGION_POLYFILL'
                 ]
             },
             LOAD_PLUGINS : {
@@ -95,6 +105,28 @@ define([
                 require('polyfills/base64');
                 resolve();
             }, 'polyfills.base64');
+        } else {
+            resolve();
+        }
+    }
+
+    function _loadVTTCuePolyfill(resolve) {
+        if (!window.VTTCue && !window.TextTrackCue) {
+            require.ensure(['polyfills/vttcue'], function(require) {
+                window.VTTCue = require('polyfills/vttcue');
+                resolve();
+            }, 'polyfills.vttcue');
+        } else {
+            resolve();
+        }
+    }
+
+    function _loadVTTRegionPolyfill(resolve) {
+        if (!window.VTTRegion) {
+            require.ensure(['polyfills/vttregion'], function(require) {
+                window.VTTRegion = require('polyfills/vttregion');
+                resolve();
+            }, 'polyfills.vttregion');
         } else {
             resolve();
         }
